Fixed Derivative Issue with Bright Autoflourescent spots
Altered the Derivative function to check for and mask extremely bright
spots in the derivative image.  Also, will now output the Median value.
Further testing is needed to determine parity with StarSearch.  Fixed
actual derivative calculation.
Further testing is needed to determine the value of the area check in
derivative function. Currently set at 100.

Minor Optimization to dots and crazypoly function calls to prevent
excessive calls to that function.

Added a check of the maxima value, so it increases if the spot count is
too high.