maximasearch uses second derivative
maximasearch calculates the change in slope of the spot count to find
when the slope levels off. Moves a window to calculate the change in
slope.
Maxima tolerance is best set around 10.

Commenting

Crazy Poly and Dots don't add bad spots (large spots) and now add to a
bad spots counter which has been added to the tables and images.