Exclusion mode added to User Defined Selection
User Defined Selection can now be set to exclude the selection, or
analyze the entire image.  User is prompted each image to set it to
exclude the selection, only analyze the selection, or analyze the entire
image.

Fixed a bug when using Filter Spots and User Defined area wouldn't work.

Fixed a crash when User Defined Area and Filter Spots were turned on and
no bright spots were detected.

Fixed a bug where an extra blank slice was added to the merged images if
no bright spots were detected