User Defined Area Change
Ran into an issue where a small user defined area would not have any
noise selected, or a very small amount selected.  Noise and Background
were being thresholded based on the entire image, not the selection.
Noise and Background are now measured relative to the selection made,
not to the entire image.